packages/chuk-virtual-expert-time/pyproject.toml
[project]
name = "chuk-virtual-expert-time"
version = "2.0.0"
description = "Time virtual expert - accurate time and timezone operations for LLM routing"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Christopher Hay" }
]
keywords = [
    "virtual-expert",
    "time",
    "timezone",
    "llm",
    "lazarus",
    "pydantic",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "chuk-virtual-expert>=2.0.0",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
mcp = ["chuk-mcp-time>=0.1.0"]
standalone = ["httpx>=0.24.0", "tzdata"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "coverage[toml]>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-python-dateutil>=2.8",
    "bandit>=1.7",
    "twine>=4.0",
    "build>=1.0",
]

[project.urls]
Homepage = "https://github.com/chrishayuk/virtual-experts"
Documentation = "https://github.com/chrishayuk/virtual-experts#readme"
Repository = "https://github.com/chrishayuk/virtual-experts"
Issues = "https://github.com/chrishayuk/virtual-experts/issues"

[project.entry-points."chuk_virtual_expert.plugins"]
time = "chuk_virtual_expert_time:TimeExpert"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chuk_virtual_expert_time"]

[tool.uv.sources]
chuk-virtual-expert = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/chuk_virtual_expert_time"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 90
show_missing = true

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]  # assert used for tests
